+++ NEWS.md
@@ -1,3 +1,17 @@
+# gimms 1.2.4.9001 (2026-03-21)
+
+#### ✨ features and improvements
+
+  * Revert {Kendall} and {zyp} removal (#16)
+  * Add initial unit tests for `significantTau()`
+
+#### 🐛 bug fixes
+
+#### 💬 documentation etc
+
+#### 🍬 miscellaneous
+
+
 # gimms 1.2.4 (2025-11-30)
 
 #### 🍬 miscellaneous